Fix variables [READ THE DESCRIPTION IF YOU'RE SELF-HOSTING]
This update will fuck up duckhunt if you're self-hosting it. You need to "migrate" the database before with the migration.py script included with this commit. WARNING : use the script at your own risk, and backup your database before. Make sure the bot is updated before starting it with the migrated database. If this is too hard for you, consider switching to the official version. There's no support for the slef-hosted bots. Also, the script is in french.