Hi Arthur,
We believe that in the mean while, we have pushed a fix for the timeout after the script has ended. Could you try and spin off a build with the workaround of `filter_secrets: false` removed?

Please do let us know if you would encounter any irregularities.

Thanks, and have a great day!

[REDACTED]